вношу правки в нейминг, исправляю ссылку
